Move AbstractRuleTestCase and TestHelper to src
