+++ tests/Rules/File/FileName/FileNameRuleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace TwigCsFixer\Tests\Rules\File\FileName;
 
 use TwigCsFixer\Rules\File\FileNameRule;
-use TwigCsFixer\Tests\Rules\AbstractRuleTestCase;
+use TwigCsFixer\Test\AbstractRuleTestCase;
 
 final class FileNameRuleTest extends AbstractRuleTestCase
 {
