+++ tests/Rules/Function/NamedArgumentSpacing/NamedArgumentSpacingRuleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace TwigCsFixer\Tests\Rules\Function\NamedArgumentSpacing;
 
 use TwigCsFixer\Rules\Function\NamedArgumentSpacingRule;
-use TwigCsFixer\Tests\Rules\AbstractRuleTestCase;
+use TwigCsFixer\Test\AbstractRuleTestCase;
 
 final class NamedArgumentSpacingRuleTest extends AbstractRuleTestCase
 {
