+++ tests/Rules/Node/ForbiddenBlock/ForbiddenBlockRuleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace TwigCsFixer\Tests\Rules\Node\ForbiddenBlock;
 
 use TwigCsFixer\Rules\Node\ForbiddenBlockRule;
-use TwigCsFixer\Tests\Rules\AbstractRuleTestCase;
+use TwigCsFixer\Test\AbstractRuleTestCase;
 
 final class ForbiddenBlockRuleTest extends AbstractRuleTestCase
 {
