+++ tests/Rules/Punctuation/TrailingCommaSingleLine/TrailingCommaSingleLineRuleTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace TwigCsFixer\Tests\Rules\Punctuation\TrailingCommaSingleLine;
 
 use TwigCsFixer\Rules\Punctuation\TrailingCommaSingleLineRule;
-use TwigCsFixer\Tests\Rules\AbstractRuleTestCase;
+use TwigCsFixer\Test\AbstractRuleTestCase;
 
 final class TrailingCommaSingleLineRuleTest extends AbstractRuleTestCase
 {
